+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this dbapi driver will be documented in this file.
 
-## Unreleased
+## 0.4.0, 2026-06-15
 
 ### Added